+++ pybann/__init__.py
@@ -1,1 +1,2 @@
 from . import activation
+from .network import Network